[chores] Moved check_email_verification to handlers.py
The file handlers.py is where we store signal receivers
in this module, let's maintain consistency.
